Minor changes made to most classes and interfaces, mostly cleanup. Fixed misspelled names and tweaked class names so that they fit naming conventions better. Made all interfaces public instead of internal, otherwise they can't be implemented. The was an issue with namespaces where inside of new folders new namespaces were made, fixed this so that all files adhere to namespace Game_Project. Fixed a few minor bugs
