+++ Game Project/Commands/PlayerMoveRightCommand.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.Text;
-using Game_Project.Sprites;
 
 namespace Game_Project
 {
