+++ Game Project/Controllers/KeyboardControllerDraft.cs
@@ -1,9 +1,7 @@
+using Microsoft.Xna.Framework.Input;
 using System;
 using System.Collections.Generic;
 using System.Text;
-using Microsoft.Xna.Framework.Input;
-using Game_Project.Sprites;
-using Game_Project.Interfaces;
 
 namespace Game_Project
 {
