--- Game Project/Interfaces/ICollidable.cs
@@ -1,11 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Game_Project.Interfaces
-{
-    internal interface ICollidable
-    {
-        public void Collide(); // Update any appropriate states for when the player collides with another object.
-    }
-}
