+++ Game Project/Interfaces/ICollideable.cs
@@ -1,0 +1,7 @@
+namespace Game_Project
+{
+    public interface ICollideable
+    {
+        public void Collide(); // Update any appropriate states for when the player collides with another object.
+    }
+}
