+++ Game Project/Interfaces/ICommand.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Game_Project
+﻿namespace Game_Project
 {
     public interface ICommand
     {
