+++ Game Project/Interfaces/IEnemyStateMachine.cs
@@ -1,10 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Game_Project.Interfaces
+﻿namespace Game_Project
 {
-    interface IEnemyStateMachine
+    public interface IEnemyStateMachine
     {
 
         //This method is left in here in case we need it later. If it is not needed, take that boi out
