+++ Game Project/Interfaces/IPlayer.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Game_Project.Interfaces
+﻿namespace Game_Project
 {
     public interface IPlayer
     {
