+++ Game Project/Interfaces/IProjectile.cs
@@ -1,7 +1,4 @@
 ﻿using Microsoft.Xna.Framework;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace Game_Project
 {
