+++ Game Project/Interfaces/ISprite.cs
@@ -1,10 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using Microsoft.Xna.Framework.Graphics;
+﻿using Microsoft.Xna.Framework.Graphics;
 using Microsoft.Xna.Framework;
 
-namespace Game_Project.Interfaces
+namespace Game_Project
 {
     public interface ISprite
     {
