--- Game Project/Interfaces/IUpdatable.cs
@@ -1,11 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Game_Project.Interfaces
-{
-    internal interface IUpdatable
-    {
-        public void Update(); // Update any appropriate states for the game.
-    }
-}
