+++ Game Project/Interfaces/IUpdateable.cs
@@ -1,0 +1,9 @@
+using Microsoft.Xna.Framework;
+
+namespace Game_Project
+{
+    public interface IUpdateable
+    {
+        public void Update(GameTime gameTime); // Update any appropriate states for the game.
+    }
+}
