+++ Game Project/Player/RightIdle.cs
@@ -1,11 +1,8 @@
-﻿using Microsoft.Xna.Framework.Graphics;
-using Microsoft.Xna.Framework;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Text;
-using Game_Project.Interfaces;
 
-namespace Game_Project.Sprites
+namespace Game_Project
 {
     class RightIdle : IPlayer
     {
