+++ Game Project/Sprites/GenericSprite.cs
@@ -3,9 +3,8 @@ using Microsoft.Xna.Framework.Graphics;
 using System;
 using System.Collections.Generic;
 using System.Text;
-using Game_Project.Interfaces;
 
-namespace Game_Project.Sprites
+namespace Game_Project
 {
     class GenericSprite : ISprite
     {
